Refactor: error 패키지 정렬 및 크롤러 예외처리 수정
